[FIX] dunia per seederan

+++ app/Http/Controllers/Api/UserController.php
@@ -98,6 +98,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        if ($user->id == 1) return response()->json(['message' => 'Admin with id 1 can not deleted!']);
         abort_if(!auth()->user()->tokenCan('user_delete'), 403);
         $user->delete();
         return $this->deletedResponse();
@@ -105,6 +106,7 @@ class UserController extends Controller
 
     public function forceDelete($id)
     {
+        if ($id == 1) return response()->json(['message' => 'Admin with id 1 can not deleted!']);
         abort_if(!auth()->user()->tokenCan('user_delete'), 403);
         $user = User::withTrashed()->findOrFail($id);
         $user->forceDelete();
